project/places/admin.py
from django.contrib import admin
from django.utils.html import format_html

from adminsortable2.admin import SortableAdminBase, SortableInlineAdminMixin

from places.models import Image, Place


class ImageInline(SortableInlineAdminMixin, admin.TabularInline):
    model = Image
    extra = 0
    readonly_fields = ['image_preview', ]
    fields = ['image', 'image_preview',]

    def image_preview(self, obj):
        return format_html(
            '<img src="{url}" style="max-width: 255px; max-height: 200px;" />',
            url=obj.image.url
        )


@admin.register(Place)
class PlaceAdmin(SortableAdminBase, admin.ModelAdmin):
    inlines = [ImageInline,]


@admin.register(Image)
class ImageAdmin(admin.ModelAdmin):
    raw_id_fields = ('place',)
    readonly_fields = ['image_preview']
