requirements.txt
pillow==9.5.0
Django==3.2.25
django-admin-sortable2==1.0.4
django-tinymce==3.6.1
environs==10.1.0
requests==2.31.0